+++ conceitos.html
@@ -9,17 +9,21 @@
 <body>
   <h1>Conceitos</h1>
 
-  <!-- <script src="sintaxe.js"></script> -->
-  <!-- <script src="paradigmas.js"></script> -->
-  <!-- <script src="objetos-literais-1.js"></script> -->
-  <!-- <script src="objetos-literais-2.js"></script> -->
-  <!-- <script src="objeto-window-global.js"></script> -->
-  <!-- <script src="uso-do-this.js"></script> -->
-  <!-- <script src="constructor-pattern.js"></script> -->
-  <!-- <script src="prototype.js"></script> -->
-  <!-- <script src="escopo.js"></script> -->
-  <!-- <script src="module-pattern.js"></script> -->
-  <!-- <script src="classe-automovel.js"></script> -->
-  <script src="televisao.js"></script>
+  <!-- <script async src="sintaxe.js"></script>
+  <script async src="paradigmas.js"></script>
+  <script async src="objetos-literais-1.js"></script>
+  <script async src="objetos-literais-2.js"></script>
+  <script async src="objeto-window-global.js"></script>
+  <script async src="uso-do-this.js"></script>
+  <script async src="constructor-pattern.js"></script>
+  <script async src="prototype.js"></script>
+  <script async src="escopo.js"></script>
+  <script async src="module-pattern.js"></script>
+  <script async src="classe-automovel.js"></script> -->
+  <script defer src="classes/endereco.js" type="module"></script>
+  <script defer src="classes/fabricante.js" type="module"></script>
+  <script defer src="classes/televisao.js" type="module"></script>
+  <script defer src="uso-de-classes.js" type="module"></script>
+  
 </body>
 </html>
